+++ changelog.md
@@ -1,4 +1,9 @@
-0.1.1 (UPCOMING):
+0.2.0 (UPCOMING):
   - Changelog.md added.
   - Configuration has been refactored, now has full test coverage.
-  - CONFIG_PARAM_ERROR added.
+  - Created new image creation module with more options.
+  - Rewrote "renderer" as mp3-to-mp4.py
+  - Rewrote config.py
+  - More user options
+  - Smoother handling of configuration files.
+